requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
bcrypt==4.0.1
cad==1.1.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2
django-cors-headers==3.14.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dynaconf==3.1.11
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
pycryptodome==3.16.0
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
redis==4.4.2
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.14
